Refactor: main페이지 바뀐 구조로 사용되도록 수정
