+++ src/component/presenter/icons/MainLogo.jsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const MainLogo = styled.div`
+  border: 1px solid #011627;
+  width: 15%;
+  height: 80%;
+`;
+
+export default MainLogo;
