+++ src/component/presenter/layouts/HeaderContainer.jsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+const HeaderContainer = styled.header`
+  display: flex;
+  justify-content: space-between;
+  align-items:center;
+  width: 100%;
+  border: 1px solid #bbb;
+  border-top: 0.4rem solid #2EC4B6
+  padding: 1%;
+  height: 15vh;
+  background-color: #EFFFE9
+`;
+
+export default HeaderContainer;
