+++ src/component/presenter/layouts/ListContainer.jsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const ListContainer = styled.div`
+  height: 85vh;
+  width: 100%;
+  margin: 0 auto;
+  overflow-y: scroll;
+`;
+
+export default ListContainer;
